% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{AcidExperiment-package}
\alias{AcidExperiment}
\alias{AcidExperiment-package}
\title{AcidExperiment}
\description{
Toolkit to extend the functionality of SummarizedExperiment.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r.acidgenomics.com/packages/acidexperiment/}
  \item \url{https://github.com/acidgenomics/r-acidexperiment/}
  \item Report bugs at \url{https://github.com/acidgenomics/r-acidexperiment/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
  \item Rory Kirchner \email{roryk@alum.mit.edu} (\href{https://orcid.org/0000-0003-4814-5885}{ORCID}) [contributor]
}

}
\keyword{internal}
