% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\name{AcidExperimentTestsUrl}
\alias{AcidExperimentTestsUrl}
\title{AcidExperiment test data URL}
\usage{
AcidExperimentTestsUrl
}
\description{
AcidExperiment test data URL
}
\note{
Updated 2023-10-03.
}
\examples{
AcidExperimentTestsUrl
}
\keyword{internal}
