% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/aggregate-methods.R
\name{aggregate}
\alias{aggregate}
\alias{aggregate,Matrix-method}
\alias{aggregate,SummarizedExperiment-method}
\alias{aggregate,matrix-method}
\title{Aggregate}
\usage{
aggregate(x, ...)

\S4method{aggregate}{Matrix}(x, by, fun = c("sum", "mean", "n"))

\S4method{aggregate}{SummarizedExperiment}(x, assay = 1L, ...)

\S4method{aggregate}{matrix}(x, by, fun = c("sum", "mean", "median", "geometricMean", "n"))
}
\arguments{
\item{x}{Object.}

\item{by}{\code{factor}.
Aggregation groupings. The new aggregate names are defined as the \code{factor}
\code{levels}, and the original, unaggregated names are defined as the \code{names}.}

\item{fun}{\code{character(1)}.
Name of the aggregation function to apply.
Uses \code{match.arg()} internally.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
}
\description{
Aggregate
}
\section{Methods (by class)}{
\itemize{
\item \code{SummarizedExperiment}: Arguments pass through to \code{matrix} or \code{Matrix} method, depending on the class
of matrix defined in requested \code{assay}.
}}

\note{
Updated 2021-09-10.
}
\examples{
counts <- matrix(
    data = c(
        0L, 2L, 2L, 2L,
        2L, 0L, 2L, 2L,
        2L, 2L, 0L, 2L,
        2L, 2L, 2L, 0L
    ),
    nrow = 4L,
    ncol = 4L,
    byrow = TRUE,
    dimnames = list(
        paste0("transcript", seq_len(4L)),
        paste(
            paste0("sample", rep(seq_len(2L), each = 2L)),
            paste0("replicate", rep(seq_len(2L), times = 2L)),
            sep = "_"
        )
    )
)

genes <- factor(paste0("gene", rep(seq_len(2L), each = 2L)))
names(genes) <- rownames(counts)
print(genes)

## matrix ====
print(counts)
aggregate(counts, by = genes)

## Matrix ====
sparse <- as(counts, "sparseMatrix")
print(sparse)
aggregate(sparse, by = genes)
}
\seealso{
\itemize{
\item \code{stats::aggregate()}.
\item \code{S4Vectors::aggregate()}.
\item \code{Matrix.utils::aggregate.Matrix()} (defunct).
\item \code{muscat::aggregateData()}.
}
}
