% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/aggregateRows-methods.R
\name{aggregateRows}
\alias{aggregateRows}
\alias{aggregateRows,Matrix-method}
\alias{aggregateRows,SummarizedExperiment-method}
\alias{aggregateRows,matrix-method}
\title{Aggregate rows}
\usage{
aggregateRows(x, ...)

\S4method{aggregateRows}{Matrix}(x, by, fun = c("sum", "mean"))

\S4method{aggregateRows}{SummarizedExperiment}(x, col = "aggregate", fun = "sum")

\S4method{aggregateRows}{matrix}(x, by, fun = c("sum", "mean", "median", "geometricMean", "n"))
}
\arguments{
\item{x}{Object.}

\item{by}{\code{factor}.
Aggregation groupings. The new aggregate names are defined as the \code{factor}
\code{levels}, and the original, unaggregated names are defined as the \code{names}.}

\item{fun}{\code{character(1)}.
Name of the aggregation function to apply.
Uses \code{match.arg()} internally.}

\item{col}{\code{character(1)}.
Name of column in either \code{rowData()} or \code{colData()} that defines the
desired aggregation groupings.}

\item{...}{Additional arguments.}
}
\value{
Modified object of same class, with rows aggregated.
}
\description{
Aggregate rows
}
\section{Methods (by class)}{

\itemize{
\item \code{matrix}, \code{Matrix}:
Aggregate using a grouping \code{factor}.
\item \code{SummarizedExperiment}:
Aggregate data slotted in \code{assays()} using an automatically generated
grouping \code{factor}, which is obtained from a user-defined column
(\code{col} argument) in either the \code{rowData()} or \code{colData()} of the object.
Slot an \code{aggregate} column into \code{rowData()} for \code{aggregateRows()}, or into
\code{colData()} for \code{aggregateCols()}. This method will define the \code{groupings}
automatically, and perform the aggregation.
}
}

\examples{
counts <- matrix(
    data = c(
        0L, 2L, 2L, 2L,
        2L, 0L, 2L, 2L,
        2L, 2L, 0L, 2L,
        2L, 2L, 2L, 0L
    ),
    nrow = 4L,
    ncol = 4L,
    byrow = TRUE,
    dimnames = list(
        paste0("transcript", seq_len(4L)),
        paste(
            paste0("sample", rep(seq_len(2L), each = 2L)),
            paste0("replicate", rep(seq_len(2L), times = 2L)),
            sep = "_"
        )
    )
)

genes <- factor(paste0("gene", rep(seq_len(2L), each = 2L)))
names(genes) <- rownames(counts)
print(genes)

## matrix ====
print(counts)
aggregateRows(counts, by = genes)

## Matrix ====
sparse <- as(counts, "sparseMatrix")
print(sparse)
aggregateRows(sparse, by = genes)

## SummarizedExperiment ====
se <- SummarizedExperiment(
    assays = SimpleList(counts = counts),
    rowData = DataFrame(aggregate = genes)
)
print(se)
aggregateRows(se)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
