% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/autopadZeros-methods.R
\name{autopadZeros}
\alias{autopadZeros}
\alias{autopadZeros,SummarizedExperiment-method}
\alias{autopadZeros,matrix-method}
\title{Autopad zeros}
\usage{
autopadZeros(object, ...)

\S4method{autopadZeros}{SummarizedExperiment}(
  object,
  rownames = FALSE,
  colnames = TRUE,
  sampleNames = TRUE,
  sort = TRUE
)

\S4method{autopadZeros}{matrix}(object, rownames = FALSE, colnames = TRUE, sort = TRUE)
}
\arguments{
\item{object}{Object.}

\item{rownames}{\code{logical(1)}.
Apply to row names.}

\item{colnames}{\code{logical(1)}.
Apply to column names.}

\item{sampleNames}{\code{logical(1)}.
\emph{Applies to \code{SummarizedExperiment} method.} #' If \code{sampleName} column is
defined in \code{colData()}, these values will also get padded, if necessary.
This improves downstream handling in functions that rely on this feature.}

\item{sort}{\code{logical(1)}.
Resort using \code{sort}.}

\item{...}{Additional arguments.}
}
\value{
\code{character}.
}
\description{
Autopad zeros
}
\details{
For methods on objects supporting \code{dim()} (e.g. \code{matrix}), the object will be
returned with the rows and/or columns resorted by default. This does not
apply to the \code{character} method defined in syntactic.
}
\note{
Updated 2021-09-02.
}
\examples{
data( RangedSummarizedExperiment, package = "AcidTest")

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
autopadZeros(object, rownames = TRUE, colnames = TRUE)
}
