% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/combine-methods.R
\name{combine}
\alias{combine}
\alias{combine,SummarizedExperiment,SummarizedExperiment-method}
\title{Combine multiple objects}
\usage{
combine(x, y, ...)

\S4method{combine}{SummarizedExperiment,SummarizedExperiment}(x, y)
}
\arguments{
\item{x}{Object.}

\item{y}{Object.}

\item{...}{Additional arguments.}
}
\value{
A single value of the same class as the input values.
}
\description{
Combine multiple objects
}
\details{
We're attempting to make this as strict as possible, requiring:
\itemize{
\item Rows (genes) across objects must be identical.
\item \link[SummarizedExperiment:RangedSummarizedExperiment-class]{rowRanges} and/or
\link[SummarizedExperiment:SummarizedExperiment-class]{rowData}
\link[S4Vectors:Annotated-class]{metadata} must be identical.
\item \link[SummarizedExperiment:SummarizedExperiment-class]{colData} must contain the same columns.
\item Specific metadata must be identical (see \code{metadata} argument).
}
}
\note{
Updated 2022-02-07.
}
\examples{
data(RangedSummarizedExperiment, package = "AcidTest")

## SummarizedExperiment ====
x <- RangedSummarizedExperiment
colnames(x) <- paste0(
    "sample",
    sprintf("\%02d", seq_len(ncol(x)))
)

y <- x
colnames(y) <- paste0(
    "sample",
    sprintf("\%02d", seq(from = ncol(y) + 1L, to = ncol(y) * 2L))
)

## Combine the two objects.
c <- combine(x, y)
sampleData(c)
print(c)
}
\seealso{
\itemize{
\item \code{help("merge.Matrix", "Matrix.utils")}.
}
}
