% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/convertSampleIdsToNames-methods.R
\name{convertSampleIdsToNames}
\alias{convertSampleIdsToNames}
\alias{convertSampleIdsToNames,SummarizedExperiment-method}
\title{Convert sample identifiers to names}
\usage{
convertSampleIdsToNames(object, ...)

\S4method{convertSampleIdsToNames}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object of same class, with valid sample IDs remapped to
human-readable sample names.
}
\description{
Convert sample identifiers to names
}
\note{
Updated 2021-02-02.
}
\examples{
data(RangedSummarizedExperiment, package = "AcidTest")

## SummarizedExperiment ===
object <- RangedSummarizedExperiment
convertSampleIdsToNames(object)
}
