% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/droplevels-methods.R
\name{droplevels}
\alias{droplevels}
\alias{droplevels,SummarizedExperiment-method}
\title{Drop unused levels from factors}
\usage{
droplevels(x, ...)

\S4method{droplevels}{SummarizedExperiment}(x)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
}
\description{
The function \code{droplevels} is used to drop unused levels from a
  \code{\link[base]{factor}} or, more commonly, from factors in a data frame.
}
\note{
Updated 2021-02-03.
}
\examples{
data(RangedSummarizedExperiment, package = "AcidTest")

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
S4Vectors::droplevels(object)
}
