% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/encode-methods.R
\name{encode}
\alias{encode}
\alias{encode,SummarizedExperiment-method}
\title{Encode using run-length encoding}
\usage{
encode(x, ...)

\S4method{encode}{SummarizedExperiment}(x)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
All \code{atomic} columns will be encoded to \code{Rle} S4 class.
}
\description{
Encode using run-length encoding
}
\note{
Updated 2021-02-02.
}
\examples{
## FIXME NEED RSE EXAMPLE HERE.

## DataFrame ====
binary <- seq(from = 0L, to = 1L)
df <- S4Vectors::DataFrame(
    a = rep(x = binary, times = 50L),
    b = rep(x = binary, each = 50L)
)
lapply(df, class)
x <- encode(df)
lapply(x, class)
}
\seealso{
\code{\link[S4Vectors:Rle-class]{S4Vectors::Rle()}}.
}
