% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Ensembl2Entrez-methods.R
\name{Ensembl2Entrez}
\alias{Ensembl2Entrez}
\alias{Entrez2Ensembl}
\alias{Ensembl2Entrez,RangedSummarizedExperiment-method}
\title{Ensembl-to-Entrez gene identifier mappings}
\usage{
Ensembl2Entrez(object, ...)

Entrez2Ensembl(object, ...)

\S4method{Ensembl2Entrez}{RangedSummarizedExperiment}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{Ensembl2Entrez}.
}
\description{
Ensembl-to-Entrez gene identifier mappings
}
\note{
Updated 2021-08-10.
}
\examples{
data(RangedSummarizedExperiment, package = "AcidTest")
rse <- RangedSummarizedExperiment

## SummarizedExperiment ====
x <- Ensembl2Entrez(rse)
print(x)
}
