% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Ensembl2Ncbi-methods.R
\name{Ensembl2Ncbi}
\alias{Ensembl2Ncbi}
\alias{Ncbi2Ensembl}
\alias{Ensembl2Ncbi,RangedSummarizedExperiment-method}
\title{Ensembl-to-NCBI gene identifier mappings}
\usage{
Ensembl2Ncbi(object, ...)

Ncbi2Ensembl(object, ...)

\S4method{Ensembl2Ncbi}{RangedSummarizedExperiment}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{Ensembl2Ncbi}.
}
\description{
Ensembl-to-NCBI gene identifier mappings
}
\note{
Updated 2023-04-27.
}
\examples{
suppressPackageStartupMessages(library(SummarizedExperiment))
data(RangedSummarizedExperiment, package = "AcidTest")

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
rowRanges(object) <- as(rowRanges(object), "EnsemblGenes")
x <- Ensembl2Ncbi(object)
print(x)
}
