% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/EnsemblToNcbi-methods.R
\name{EnsemblToNcbi}
\alias{EnsemblToNcbi}
\alias{NcbiToEnsembl}
\alias{EnsemblToNcbi,RangedSummarizedExperiment-method}
\title{Ensembl-to-NCBI gene identifier mappings}
\usage{
EnsemblToNcbi(object, ...)

NcbiToEnsembl(object, ...)

\S4method{EnsemblToNcbi}{RangedSummarizedExperiment}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{EnsemblToNcbi}.
}
\description{
Ensembl-to-NCBI gene identifier mappings
}
\note{
Updated 2023-04-27.
}
\examples{
suppressPackageStartupMessages(library(SummarizedExperiment))
data(RangedSummarizedExperiment, package = "AcidTest")

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
rowRanges(object) <- as(rowRanges(object), "EnsemblGenes")
x <- EnsemblToNcbi(object)
print(x)
}
