% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/export-methods.R
\name{export}
\alias{export}
\alias{export,SummarizedExperiment-method}
\title{Export}
\usage{
export(object, ...)

\S4method{export}{SummarizedExperiment}(
  object,
  name = NULL,
  dir = getOption("acid.export.dir", default = "."),
  compress = getOption("acid.export.compress", default = FALSE),
  overwrite = getOption("acid.overwrite", default = TRUE),
  quiet = getOption("acid.quiet", default = FALSE)
)
}
\arguments{
\item{object}{Object.
An object supporting \code{\link[base:dim]{dim()}}, or a supported class capable
of being coerced to \code{data.frame}, to be written to disk.}

\item{name}{\code{character(1)}.
Name to use on disk. If \code{NULL}, will use the name of the object instead.}

\item{dir}{\code{character(1)}.
Directory path.}

\item{compress}{\code{logical(1)}.
Apply gzip compression to all files.}

\item{overwrite}{\code{logical(1)}.
Overwrite existing file on disk.}

\item{quiet}{\code{logical(1)}.
Perform command quietly, suppressing messages.}

\item{ext}{\code{character(1)}.
Output file format extension.

\code{matrix} supported arguments:
\itemize{
\item Comma separated values (CSV):
\code{"csv"}, \code{"csv.bz2"}, \code{"csv.gz"}, \code{"csv.xz"}, \code{"csv.zip"}.
\item Tab separated values (TSV):
\code{"tsv"}, \code{"tsv.bz2"}, \code{"tsv.gz"}, \code{"tsv.xz"}, \code{"tsv.zip"}.
}

\code{Matrix} (\code{sparseMatrix}) supported arguments:
\itemize{
\item MatrixMarket exchange (MTX):
\code{"mtx"}, \code{"mtx.bz2"}, \code{"mtx.gz"}, \code{"mtx.xz"}, \code{"mtx.zip"}.
}}

\item{file}{\code{character(1)}.
File path. When left unset (default), the \code{ext} and \code{dir} arguments will
be used instead.}

\item{append}{\code{logical(1)}.
Append to output file.
When enabled, automatically sets \code{overwrite} argument to \code{FALSE}.
Requires readr package to be installed.}

\item{...}{Additional arguments.}
}
\value{
Invisibily return file paths, either as \code{list} or \code{character}.
}
\description{
Export data out of R and write to disk.
}
\note{
Updated 2021-02-01.
}
\examples{
print("FIXME")
}
