% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/headtail-methods.R
\name{headtail}
\alias{headtail}
\alias{headtail,DFrame-method}
\alias{headtail,GRanges-method}
\alias{headtail,Matrix-method}
\alias{headtail,SummarizedExperiment-method}
\title{Return the first and last parts of an object}
\usage{
headtail(x, ...)

\S4method{headtail}{DFrame}(x, n = 2L)

\S4method{headtail}{GRanges}(x, n = 2L)

\S4method{headtail}{Matrix}(x, n = 2L)

\S4method{headtail}{SummarizedExperiment}(x, n = 2L)
}
\arguments{
\item{x}{Object.}

\item{n}{\code{integer(1)}.
Positive integer denoting the number of first and last items to include.}

\item{...}{Additional arguments.}
}
\value{
Invisible \code{NULL}.
}
\description{
Applies to both rows and columns, enabling quick inspection during
interactive use.
}
\details{
Inspired by the \code{\link[base:print]{print()}} method for \code{DataFrame} class
objects.
}
\section{Functions}{
\itemize{
\item \code{headtail(DFrame)}: Same method as \code{data.frame}.

\item \code{headtail(GRanges)}: Summarize the ranges.

\item \code{headtail(Matrix)}: Same method as \code{matrix}.

\item \code{headtail(SummarizedExperiment)}: Summarize the primary \code{assay}.

}}
\note{
Updated 2020-10-07.
}
\examples{
data(RangedSummarizedExperiment, package = "AcidTest")

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
headtail(object)
}
\seealso{
\itemize{
\item \code{utils::head()}, \code{utils::tail()}.
\item \code{cat()}.
}
}
