% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/metrics-methods.R
\name{metrics}
\alias{metrics}
\alias{metrics,SummarizedExperiment-method}
\title{Quality control metrics}
\usage{
metrics(object, ...)

\S4method{metrics}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{DataFrame}.
}
\description{
Quality control metrics
}
\details{
\code{metrics()} takes data stored in \code{colData()} and consistently returns a
\code{DataFrame} with \code{sampleName} and \code{interestingGroups} columns, even when
these columns are not defined in \code{colData()}. This is designed to integrate
with plotting functions that use ggplot2 internally.

Column names are always returned formatted in strict lower camel case.

This function will error intentionally if no numeric columns are defined
in \code{colData()}.
}
\section{Functions}{
\itemize{
\item \code{metrics,SummarizedExperiment-method}: Sample-level metrics.
}}

\note{
Updated 2022-05-04.
}
\examples{
data(RangedSummarizedExperiment, package = "AcidTest")

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
object <- calculateMetrics(object)
x <- metrics(object)
print(x)
}
\author{
Michael Steinbaugh, Rory Kirchner
}
