% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/nonzeroRowsAndCols-methods.R
\name{nonzeroRowsAndCols}
\alias{nonzeroRowsAndCols}
\alias{nonzeroRowsAndCols,Matrix-method}
\alias{nonzeroRowsAndCols,SummarizedExperiment-method}
\alias{nonzeroRowsAndCols,matrix-method}
\title{Subset object to keep only non-zero rows and columns}
\usage{
nonzeroRowsAndCols(object, ...)

\S4method{nonzeroRowsAndCols}{Matrix}(object, quiet = FALSE)

\S4method{nonzeroRowsAndCols}{SummarizedExperiment}(object, assay = 1L, quiet = FALSE)

\S4method{nonzeroRowsAndCols}{matrix}(object, quiet = FALSE)
}
\arguments{
\item{object}{Object.}

\item{quiet}{\code{logical(1)}.
Perform command quietly, suppressing messages.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
}
\description{
Subset object to keep only non-zero rows and columns
}
\note{
Updated 2021-02-02.
}
\examples{
data(RangedSummarizedExperiment, package = "AcidTest")

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
dim(object)
x <- nonzeroRowsAndCols(object)
dim(x)
}
