% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/organism-methods.R
\name{organism}
\alias{organism}
\alias{organism,SummarizedExperiment-method}
\title{Organism}
\usage{
organism(object)

\S4method{organism}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{character(1)}.
}
\description{
Get or set the organism (i.e. species) of an object.
}
\note{
Updated 2021-02-03.

Assignment method is defined in AcidGenomes.
}
\examples{
data(RangedSummarizedExperiment, package = "AcidTest")

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
organism(object)
organism(object) <- "xxx"
organism(object)
}
