% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitizeSampleData.R
\name{sanitizeSampleData}
\alias{sanitizeSampleData}
\title{Sanitize sample data}
\usage{
sanitizeSampleData(object)
}
\arguments{
\item{object}{\code{DataFrame} (recommended) or \code{data.frame} (legacy).
Note that legacy \code{data.frame} support will be removed in a future update.}
}
\value{
\code{DataFrame}.
}
\description{
Drop any columns defined in denylist, and ensure \code{character} columns
containing any duplicate values are automatically coerced to \code{factor}.
}
\note{
Updated 2022-10-24.
}
\section{Denylist}{


Here's the current column denylist:
\itemize{
\item \code{interestingGroups}.
\item \code{sampleId}.
}
}

\examples{
data(RangedSummarizedExperiment, package = "AcidTest")
rse <- RangedSummarizedExperiment

## SummarizedExperiment ====
from <- sampleData(rse)
print(from)
to <- sanitizeSampleData(from)
all(vapply(to, is.factor, logical(1L)))
print(to)
}
