% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/sizeFactors-methods.R
\name{sizeFactors}
\alias{sizeFactors}
\alias{sizeFactors<-}
\alias{sizeFactors,SummarizedExperiment-method}
\alias{sizeFactors<-,SummarizedExperiment-method}
\title{Size factors}
\usage{
sizeFactors(object, ...)

sizeFactors(object, ...) <- value

\S4method{sizeFactors}{SummarizedExperiment}(object)

\S4method{sizeFactors}{SummarizedExperiment}(object) <- value
}
\arguments{
\item{object}{Object.}

\item{value}{Value to assign.}

\item{...}{Additional arguments.}
}
\value{
\code{numeric}.
Names correspond to object column names.
}
\description{
Size factors
}
\note{
Updated 2021-02-02.
}
\examples{
data(RangedSummarizedExperiment, package = "AcidTest")

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
object <- estimateSizeFactors(object)
head(sizeFactors(object))
mean(sizeFactors(object))
}
\seealso{
\itemize{
\item \code{DESeq2::sizeFactors()}.
\item \code{DESeq2::estimateSizeFactors()}.
\item \code{DESeq2::estimateSizeFactorsForMatrix()}.
}
}
