% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripTranscriptVersions-methods.R
\name{stripTranscriptVersions}
\alias{stripTranscriptVersions}
\alias{stripTranscriptVersions,SummarizedExperiment-method}
\title{Strip transcript identifier versions}
\usage{
\S4method{stripTranscriptVersions}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Contains transcript identifiers without version numbers.
}
\description{
Strip transcript identifier versions
}
\details{
Intentionally does not strip gene versions, if defined in the object.
To perform that action in parallel, use \code{stripVersions} instead.
}
\note{
Updated 2021-08-09.
}
\examples{
data(SummarizedExperiment_transcripts, package = "AcidTest")
rowData <- SummarizedExperiment::rowData

## SummarizedExperiment ====
object <- SummarizedExperiment_transcripts
head(rowData(object)[["txId"]])
head(rowData(object)[["geneId"]])
head(rownames(object))
object <- stripGeneVersions(object)
object <- stripTranscriptVersions(object)
head(rowData(object)[["txId"]])
head(rowData(object)[["geneId"]])
head(rownames(object))
}
\seealso{
\itemize{
\item http://www.ensembl.org/info/genome/stable_ids/index.html
}
}
