% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/stripVersions-methods.R
\name{stripGeneVersions}
\alias{stripGeneVersions}
\alias{stripTranscriptVersions}
\alias{stripVersions}
\alias{stripGeneVersions,SummarizedExperiment-method}
\alias{stripTranscriptVersions,SummarizedExperiment-method}
\title{Strip transcript identifier versions}
\usage{
stripGeneVersions(object, ...)

stripTranscriptVersions(object, ...)

\S4method{stripGeneVersions}{SummarizedExperiment}(object)

\S4method{stripTranscriptVersions}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Contains transcript identifiers without version numbers.
}
\description{
Strip transcript identifier versions
}
\details{
Intentionally does not strip gene versions, if defined in the object.
To perform that action in parallel, use \code{stripVersions} instead.
}
\note{
Updated 2021-02-03.
}
\examples{
data(SummarizedExperiment_transcripts, package = "AcidTest")

## SummarizedExperiment ====
object <- SummarizedExperiment_transcripts
head(SummarizedExperiment::rowData(object)[["txId"]])
head(SummarizedExperiment::rowData(object)[["geneId"]])
head(rownames(object))
object <- stripGeneVersions(object)
object <- stripTranscriptVersions(object)
head(SummarizedExperiment::rowData(object)[["txId"]])
head(SummarizedExperiment::rowData(object)[["geneId"]])
head(rownames(object))
}
\seealso{
\itemize{
\item http://www.ensembl.org/info/genome/stable_ids/index.html
}
}
