% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Tx2Gene-methods.R
\name{Tx2Gene}
\alias{Tx2Gene}
\alias{Tx2Gene,SummarizedExperiment-method}
\title{Transcript-to-gene identifier mappings}
\usage{
Tx2Gene(object, ...)

\S4method{Tx2Gene}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{Tx2Gene}.
}
\description{
Transcript-to-gene identifier mappings
}
\note{
Updated 2021-02-03.
}
\examples{
data(SummarizedExperiment_transcripts, package = "AcidTest")

## SummarizedExperiment ====
object <- SummarizedExperiment_transcripts
t2g <- Tx2Gene(object)
print(t2g)
}
\seealso{
\code{\link[AcidGenomes:makeTx2Gene]{makeTx2Gene()}}.
}
