% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/TxToGene-methods.R
\name{TxToGene}
\alias{TxToGene}
\alias{TxToGene,SummarizedExperiment-method}
\title{Transcript-to-gene mappings}
\usage{
TxToGene(object, ...)

\S4method{TxToGene}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{TxToGene}.
}
\description{
Transcript-to-gene mappings
}
\note{
Updated 2021-02-03.
}
\examples{
data(SummarizedExperiment_transcripts, package = "AcidTest")

## SummarizedExperiment ====
object <- SummarizedExperiment_transcripts
t2g <- TxToGene(object)
print(t2g)
}
\seealso{
\code{\link[AcidGenomes:makeTxToGene]{makeTxToGene()}}.
}
