% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGlobals.R
\docType{data}
\name{AcidExperimentTestsUrl}
\alias{AcidExperimentTestsUrl}
\title{AcidExperiment test data URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
AcidExperimentTestsUrl
}
\description{
AcidExperiment test data URL
}
\note{
Updated 2021-02-03.
}
\examples{
AcidExperimentTestsUrl
}
\keyword{internal}
