% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calculateMetrics-methods.R
\name{calculateMetrics}
\alias{calculateMetrics}
\alias{calculateMetrics,Matrix-method}
\alias{calculateMetrics,SummarizedExperiment-method}
\alias{calculateMetrics,matrix-method}
\title{Calculate quality control metrics}
\usage{
calculateMetrics(object, ...)

\S4method{calculateMetrics}{Matrix}(object, rowData = NULL, prefilter = FALSE)

\S4method{calculateMetrics}{SummarizedExperiment}(object, assay = 1L, prefilter = FALSE)

\S4method{calculateMetrics}{matrix}(object, rowData = NULL, prefilter = FALSE)
}
\arguments{
\item{object}{Object.}

\item{rowData}{\code{DataFrame}.
Metadata describing the assay rows, if genomic ranges are not available.
\emph{Use \code{rowRanges} (\code{GenomicRanges}) instead, if possible}.}

\item{prefilter}{\code{logical(1)}.
Drop very low quality samples/cells from the object.
This can resize the number of columns but the rows (i.e. features) do not
change with this operation.}

\item{assay}{\code{vector(1)}.
Assay name or index position.}

\item{...}{Additional arguments.}
}
\value{
\itemize{
\item \code{matrix} / \code{Matrix}: \code{DFrame} containing metrics.
\item \code{SummarizedExperiment}: Modified object, with metrics in
\code{colData()}.
}
}
\description{
Calculate quality control metrics
}
\details{
Input a raw count matrix. Do not use size factor adjusted or log normalized
counts here.
}
\note{
Updated 2021-09-11.
}
\examples{
data(RangedSummarizedExperiment, package = "AcidTest")

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
names(SummarizedExperiment::colData(object))
object <- calculateMetrics(object)
names(SummarizedExperiment::colData(object))
}
\author{
Michael Steinbaugh, Rory Kirchner
}
