% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/calculateMetrics-methods.R
\name{calculateMetrics}
\alias{calculateMetrics}
\alias{calculateMetrics,matrix-method}
\alias{calculateMetrics,Matrix-method}
\alias{calculateMetrics,RangedSummarizedExperiment-method}
\title{Calculate quality control metrics}
\usage{
calculateMetrics(object, ...)

\S4method{calculateMetrics}{matrix}(object, rowRanges = NULL, prefilter = FALSE)

\S4method{calculateMetrics}{Matrix}(object, rowRanges = NULL, prefilter = FALSE)

\S4method{calculateMetrics}{RangedSummarizedExperiment}(object, prefilter = FALSE)
}
\arguments{
\item{object}{Object.}

\item{rowRanges}{\code{GRanges} or \code{GRangesList}.
Genomic ranges (e.g. genome annotations). Metadata describing the assay
rows.}

\item{prefilter}{\code{logical(1)}.
Drop very low quality samples/cells from the object.
This can resize the number of columns but the rows (i.e. features) do not
change with this operation.}

\item{...}{Additional arguments.}
}
\value{
\itemize{
\item \code{matrix} / \code{Matrix}: \code{DataFrame} containing metrics.
\item \code{SummarizedExperiment}: Modified object, with metrics in
\code{\link[SummarizedExperiment:SummarizedExperiment-class]{colData()}}.
}
}
\description{
Calculate quality control metrics
}
\details{
Input a raw count matrix. Do not use size factor adjusted or log normalized
counts here.
}
\note{
Updated 2020-02-03.
}
\examples{
data(RangedSummarizedExperiment, package = "AcidTest")

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
names(colData(object))
object <- calculateMetrics(object)
names(colData(object))
}
\author{
Michael Steinbaugh, Rory Kirchner
}
