% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/convertGenesToSymbols-methods.R
\name{convertGenesToSymbols}
\alias{convertGenesToSymbols}
\alias{convertSymbolsToGenes}
\alias{convertGenesToSymbols,SummarizedExperiment-method}
\alias{convertSymbolsToGenes,SummarizedExperiment-method}
\title{Convert genes to symbols}
\usage{
convertGenesToSymbols(object, ...)

convertSymbolsToGenes(object, ...)

\S4method{convertGenesToSymbols}{SummarizedExperiment}(object, strict = FALSE)

\S4method{convertSymbolsToGenes}{SummarizedExperiment}(object, strict = FALSE)
}
\arguments{
\item{object}{Object.}

\item{strict}{\code{logical(1)}.
Require that all identifiers contain gene name (symbol) metadata stored
in the object. Disabled by default, to support objects containing
custom gene identifiers, such as FASTA spike-ins.}

\item{...}{Additional arguments.}
}
\value{
Modified object of same class.
}
\description{
Convert genes to symbols
}
\note{
Updated 2021-10-21.
}
\examples{
data(RangedSummarizedExperiment, package = "AcidTest")

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
x <- convertGenesToSymbols(object)
print(x)
## Interconvert back to gene IDs.
y <- convertSymbolsToGenes(x)
print(y)
}
