% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/decode-methods.R
\name{decode}
\alias{decode}
\alias{decode,SummarizedExperiment-method}
\title{Decode column data that uses run-length encoding}
\usage{
decode(x, ...)

\S4method{decode}{SummarizedExperiment}(x)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Columns will be decoded and no longer \code{Rle} class.
}
\description{
The Rle class is a general container for storing an atomic vector
  that is stored in a run-length encoding format. It is based on the
  \code{\link[base]{rle}} function from the base package.
}
\note{
Updated 2021-02-03.
}
\examples{
data(RangedSummarizedExperiment, package = "AcidTest")

## SummarizedExperiment ====
## This works on rowData and colData.
object <- RangedSummarizedExperiment
object <- decode(object)
lapply(SummarizedExperiment::rowData(object), class)
}
\seealso{
\code{\link[S4Vectors:Rle-class]{S4Vectors::decode()}}.
}
