% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/encode-methods.R
\name{encode}
\alias{encode}
\alias{encode,SummarizedExperiment-method}
\title{Encode using run-length encoding}
\usage{
encode(x, ...)

\S4method{encode}{SummarizedExperiment}(x)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
All \code{atomic} columns will be encoded to \code{Rle} S4 class.
}
\description{
Encode using run-length encoding
}
\note{
Updated 2021-02-03.
}
\examples{
data(RangedSummarizedExperiment, package = "AcidTest")

## SummarizedExperiment ====
## This works on rowData and colData.
object <- RangedSummarizedExperiment
object <- encode(object)
lapply(rowData(object), class)
}
\seealso{
\code{S4Vectors::Rle()}.
}
