% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ensembl2Entrez-methods.R
\name{Ensembl2Entrez}
\alias{Ensembl2Entrez}
\alias{Ensembl2Entrez,RangedSummarizedExperiment-method}
\title{Ensembl-to-Entrez gene identifier mappings}
\usage{
\S4method{Ensembl2Entrez}{RangedSummarizedExperiment}(object, format = c("1:1", "long"))
}
\arguments{
\item{object}{Object.}
}
\value{
\code{Ensembl2Entrez}.
}
\description{
Ensembl-to-Entrez gene identifier mappings
}
\note{
Updated 2021-02-02.
}
\examples{
data(RangedSummarizedExperiment, package = "AcidTest")
rse <- RangedSummarizedExperiment

## SummarizedExperiment ====
x <- Ensembl2Entrez(rse)
print(x)
}
