% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Ensembl2Entrez-methods.R
\name{Ensembl2Entrez}
\alias{Ensembl2Entrez}
\alias{Entrez2Ensembl}
\alias{Ensembl2Entrez,RangedSummarizedExperiment-method}
\title{Ensembl-to-Entrez gene identifier mappings}
\usage{
Ensembl2Entrez(object, ...)

Entrez2Ensembl(object, ...)

\S4method{Ensembl2Entrez}{RangedSummarizedExperiment}(object, format = c("1:1", "long"))
}
\arguments{
\item{object}{Object.}

\item{format}{\code{character(1)}.
Formatting method to apply:
\itemize{
\item \code{"1:1"}: \emph{Recommended.}
Return with 1:1 mappings. For Ensembl genes that don't map 1:1 with
Entrez, pick the oldest Entrez identifier. Genes that don't map to
Entrez will contain \code{NA} in \code{entrezId} column.
\item \code{"long"}:
Return 1:many in long format.
}}

\item{...}{Additional arguments.}
}
\value{
\code{Ensembl2Entrez}.
}
\description{
Ensembl-to-Entrez gene identifier mappings
}
\note{
Updated 2021-02-03.
}
\examples{
data(RangedSummarizedExperiment, package = "AcidTest")
rse <- RangedSummarizedExperiment

## SummarizedExperiment ====
x <- Ensembl2Entrez(rse)
print(x)
}
