% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/export-methods.R
\name{export}
\alias{export}
\alias{export,SummarizedExperiment-method}
\title{Export}
\usage{
export(object, ...)

\S4method{export}{SummarizedExperiment}(
  object,
  name = NULL,
  dir = getOption("acid.export.dir", default = "."),
  compress = getOption("acid.export.compress", default = FALSE),
  overwrite = getOption("acid.overwrite", default = TRUE),
  quiet = getOption("acid.quiet", default = FALSE)
)
}
\arguments{
\item{object}{Object.
An object supporting \code{\link[base:dim]{dim()}}, or a supported class capable
of being coerced to \code{data.frame}, to be written to disk.}

\item{name}{\code{character(1)}.
Name to use on disk. If \code{NULL}, will use the name of the object instead.}

\item{dir}{\code{character(1)}.
Directory path.}

\item{compress}{\code{logical(1)}.
Apply gzip compression to all files.}

\item{overwrite}{\code{logical(1)}.
Overwrite existing file on disk.}

\item{quiet}{\code{logical(1)}.
Perform command quietly, suppressing messages.}

\item{...}{Additional arguments.}
}
\value{
Invisible \code{character}.
File path(s).
}
\description{
Export data out of R and write to disk.
}
\note{
Updated 2021-02-03.
}
\section{Row names}{



Some export utilities in R have a tendency to drop row names when writing to
disk in CSV format. For example, the \href{https://readr.tidyverse.org/}{readr} family of functions never
write row names by design. This is a \emph{really poor} default setting for
handling genomic data, which often contain gene identifiers in the row names.
Here we're performing any internal tibble coercion step to ensure row names
are always moved to a \code{"rowname"} column in the CSV export.

}

\section{Debugging}{



Note that this function currently wraps \code{vroom::voom_write()} by default
for exporting \code{data.frame} and \code{matrix} class objects.

}

\examples{
print("FIXME")
}
\seealso{
Packages:
\itemize{
\item \href{http://r-datatable.com/}{data.table}.
\item \href{http://readr.tidyverse.org}{readr}.
\item \href{https://cran.r-project.org/package=rio}{rio}.
\item \href{http://bioconductor.org/packages/rtracklayer/}{rtracklayer}.
\item \href{https://vroom.r-lib.org}{vroom}.
}

Export functions:
\itemize{
\item \code{data.table::fwrite()}.
\item \code{readr::write_csv()}.
\item \code{rio::export()}.
\item \code{rtracklayer::export()}.
\item \code{vroom::vroom_write()}.
}
}
