% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Gene2Symbol-methods.R
\name{Gene2Symbol}
\alias{Gene2Symbol}
\alias{Gene2Symbol,SummarizedExperiment-method}
\title{Gene-to-symbol mappings}
\usage{
Gene2Symbol(object, ...)

\S4method{Gene2Symbol}{SummarizedExperiment}(object, ...)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
\code{Gene2Symbol}.
}
\description{
Gene-to-symbol mappings
}
\note{
Updated 2021-08-09.
}
\examples{
data(RangedSummarizedExperiment, package = "AcidTest")

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
x <- Gene2Symbol(object)
print(x)
}
\seealso{
\code{\link[AcidGenomes:makeGene2Symbol]{makeGene2Symbol()}}.
}
