% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/stripGeneVersions-methods.R
\name{stripGeneVersions}
\alias{stripGeneVersions}
\alias{stripGeneVersions,SummarizedExperiment-method}
\title{Strip gene identifier versions}
\usage{
stripGeneVersions(object, ...)

\S4method{stripGeneVersions}{SummarizedExperiment}(object)
}
\arguments{
\item{object}{Object.}

\item{...}{Additional arguments.}
}
\value{
Modified object.
Contains gene identifiers without version numbers.
}
\description{
Strip gene identifier versions
}
\details{
Intentionally does not strip transcript versions, if defined in the object.
To perform that action in parallel, use \code{stripVersions} instead.
}
\note{
Updated 2021-08-09.
}
\examples{
data(RangedSummarizedExperiment, package = "AcidTest")
rowData <- SummarizedExperiment::rowData

## SummarizedExperiment ====
object <- RangedSummarizedExperiment
head(rowData(object)[["geneId"]])
rownames(object) <- as.character(rowData(object)[["geneId"]])
head(rownames(object))
object <- stripGeneVersions(object)
head(rowData(object)[["geneId"]])
head(rownames(object))
}
\seealso{
\itemize{
\item http://www.ensembl.org/info/genome/stable_ids/index.html
}
}
